Add metric file logging for responseTime
metrics oriented logging will be saved into `metrics.out` and then parsed be an
external program like a python script to create graphs/visualizations of the
data.

The lines in the metrics log should follow the form:
```
<n>:<d>:<c> <METRIC_NAME> <VALUE>
```

the arguments `n`,`d`,`c` follow from the project documentation as:
1. number of nodes
2. inter-request delay time
3. cs-execution time

`METRIC_NAME` should be one of the following:
1. messageComplexity
2. reponseTime
3. systemThroughput

The saved `VALUE` fields can have either the average or a single instance value
for the execution of a single node, but when the data is compiled in the final
step, all values with the same (n,d,c,METRIC_NAME) tuple will be averaged into
one value.