Changes to metric logging setup
logging now writes to separate files inside a `metrics` folder, similar to the
critical section event logger. The recorded values are all averages over the
number of messages generated.

every log file inside of the metrics directory will be read at the same time or
joined into a single file so that node id's are an unrelated to the computation
of points for graphing/charting purposes.